Rename Status notification and use Log component for text
