+++ packages/react/src/components/message/StatusNotification/StatusNotification.stories.tsx
@@ -2,11 +2,11 @@ import React from "react";
 
 import Icon from "../../asorted/Icon";
 import FlexBox from "../../layout/Flex";
-import MajorNotification, { Props } from "./index";
+import StatusNotification, { Props } from "./index";
 
 export default {
-  title: "Messages/Major Notifications",
-  component: MajorNotification,
+  title: "Messages/Status Notifications",
+  component: StatusNotification,
   argTypes: {
     text: {
       type: "string",
@@ -24,7 +24,7 @@ export default {
   },
 };
 
-export function MajorNotifications(args: Props & { active: boolean }): JSX.Element {
+export function StatusNotifications(args: Props & { active: boolean }): JSX.Element {
   const { ...props } = args;
 
   const warningBadge = <Icon name="Warning" size={50} color="palette.error.c100" weight="Light" />;
@@ -33,10 +33,10 @@ export function MajorNotifications(args: Props & { active: boolean }): JSX.Eleme
   return (
     <FlexBox justifyContent="center">
       <FlexBox width="502px" flexDirection="column" rowGap={5}>
-        <MajorNotification {...props} badge={warningBadge} hasBorder />
-        <MajorNotification {...props} badge={infoBadge} />
-        <MajorNotification {...props} badge={warningBadge} hasBorder />
-        <MajorNotification {...props} badge={infoBadge} />
+        <StatusNotification {...props} badge={warningBadge} hasBorder />
+        <StatusNotification {...props} badge={infoBadge} />
+        <StatusNotification {...props} badge={warningBadge} hasBorder />
+        <StatusNotification {...props} badge={infoBadge} />
       </FlexBox>
     </FlexBox>
   );
