+++ packages/react/src/components/message/index.ts
@@ -4,4 +4,4 @@ export { default as Notification } from "./Notification";
 export { default as Badge } from "./Notification/Badge";
 export { default as Tooltip } from "./Tooltip";
 export { default as Tip } from "./Tip";
-export { default as MajorNotification } from "./MajorNotification";
+export { default as StatusNotification } from "./StatusNotification";
